feat: factor arch-specific methods into separate modules (with macros, need some advice there)
