+++ src/avx2/mod.rs
@@ -1,3 +1,4 @@
 pub mod deser;
 pub mod stage1;
 pub mod utf8check;
+pub mod generator;
