+++ src/neon/mod.rs
@@ -1,6 +1,7 @@
 pub mod deser;
 pub mod stage1;
 pub mod utf8check;
+pub mod generator;
 mod simd;
 mod simd_llvm;
 mod intrinsics;
